Backend/Booking_Backend/Room_Booking/permissions.py
from rest_framework import permissions

class IsOwnerOrReadOnly (permissions.BasePermission):
    
    def has_object_permission(self, request, view, obj):

        def has_object_permission (self, request, view, obj):
            if request.method in permissions.SAFE_METHODS:
                return True
            
        return super().has_object_permission(request, view, obj)
    
class IsAdminOrReadOnly (permissions.BasePermission):
    def has_permission(self, request, view):
        if request.method in permissions.SAFE_METHODS:
            return True

        return request.user.is_superuser
    
    def has_object_permissions (self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
        
        return request.user.is_superuser
    
    
