Backend/Booking_Backend/requirements.txt
asgiref==3.10.0
blinker==1.9.0
boto3==1.40.30
botocore==1.40.30
build==1.3.0
certifi==2024.12.14
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.41.0
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pillow==12.1.0
psycopg==3.3.1
psycopg-binary==3.3.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
